Move BAT pipeline orchestration out of CLI
