+++ app/sources/bat/cli.py
@@ -1,48 +1,14 @@
 import argparse
 import logging
-from dataclasses import dataclass
 from datetime import date
 
-from bs4 import BeautifulSoup
 from dotenv import load_dotenv
 
-from app.sources.bat.discovery import (
-    discover_completed_auctions,
-    load_pending_discovered_listings,
-    mark_discovered_listing_handled,
-)
-from app.sources.bat.ingest import (
-    evaluate_listing_eligibility,
-    fetch_listing_html,
-    save_listing_html,
-)
-from app.sources.bat.load import load_listing
-from app.sources.bat.transform import (
-    load_pending_raw_listing_html,
-    transform_listing_html,
-)
+from app.pipeline import bat as bat_pipeline
 
 load_dotenv()
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class BatchIngestSummary:
-    selected: int = 0
-    scrape_attempted: int = 0
-    scrape_failed: int = 0
-    rejected: int = 0
-    raw_html_stored: int = 0
-    accepted: int = 0
-
-
-@dataclass
-class BatchTransformSummary:
-    selected: int = 0
-    transformed_and_loaded: int = 0
-    transform_failed: int = 0
-    load_failed: int = 0
 
 
 def build_parser():
@@ -77,98 +43,6 @@ def configure_logging():
     )
 
 
-def ingest_listing(listing_id):
-    html = fetch_listing_html(listing_id)
-    save_listing_html(listing_id, html)
-
-
-def transform_listing(listing_id):
-    transformed_listing = transform_listing_html(listing_id)
-    load_listing(transformed_listing)
-
-
-def run_listing(listing_id):
-    ingest_listing(listing_id)
-    transform_listing(listing_id)
-
-
-def discover_listings(scrape_date, max_candidates=None):
-    return discover_completed_auctions(
-        scrape_date=scrape_date,
-        max_candidates=max_candidates,
-    )
-
-
-def ingest_discovered_listings(batch_size=None):
-    summary = BatchIngestSummary()
-    pending_rows = load_pending_discovered_listings(limit=batch_size)
-
-    for row in pending_rows:
-        summary.selected += 1
-        listing_id = row["source_listing_id"]
-
-        summary.scrape_attempted += 1
-        try:
-            html = fetch_listing_html(listing_id)
-        except Exception:
-            summary.scrape_failed += 1
-            logger.error("BAT ingest-discovered scrape failed for listing_id=%s", listing_id)
-            continue
-
-        soup = BeautifulSoup(html, "html.parser")
-        eligible, reason = evaluate_listing_eligibility(soup, listing_id)
-        mark_discovered_listing_handled(listing_id, eligible, reason)
-        if not eligible:
-            logger.info(
-                "BAT ingest-discovered listing rejected for listing_id=%s reason=%s",
-                listing_id,
-                reason,
-            )
-            summary.rejected += 1
-            continue
-
-        save_listing_html(listing_id, html, url=row["url"])
-        summary.raw_html_stored += 1
-        summary.accepted += 1
-
-    return summary
-
-
-def transform_discovered_listings(batch_size=None):
-    summary = BatchTransformSummary()
-    pending_rows = load_pending_raw_listing_html(limit=batch_size)
-
-    for row in pending_rows:
-        summary.selected += 1
-        listing_id = row["source_listing_id"]
-
-        try:
-            transformed_listing = transform_listing_html(listing_id)
-        except Exception as exc:
-            summary.transform_failed += 1
-            logger.error(
-                "BAT transform-discovered row failed for listing_id=%s stage=transform error=%s",
-                listing_id,
-                exc,
-            )
-            continue
-
-        try:
-            load_listing(transformed_listing)
-        except Exception as exc:
-            summary.load_failed += 1
-            logger.error(
-                "BAT transform-discovered row failed for listing_id=%s stage=load error=%s",
-                listing_id,
-                exc,
-            )
-            continue
-
-        summary.transformed_and_loaded += 1
-
-    return summary
-
-
 def main(argv=None):
     configure_logging()
     args = build_parser().parse_args(argv)
@@ -180,7 +54,7 @@ def main(argv=None):
                 args.scrape_date.isoformat(),
                 args.max_candidates,
             )
-            summary = discover_listings(
+            summary = bat_pipeline.discover_listings(
                 scrape_date=args.scrape_date,
                 max_candidates=args.max_candidates,
             )
@@ -208,7 +82,7 @@ def main(argv=None):
                 "BAT ingest-discovered command started for batch_size=%s",
                 args.batch_size,
             )
-            summary = ingest_discovered_listings(batch_size=args.batch_size)
+            summary = bat_pipeline.ingest_discovered_listings(batch_size=args.batch_size)
             logger.info(
                 "BAT ingest-discovered summary selected=%s scrape_attempted=%s scrape_failed=%s rejected=%s raw_html_stored=%s accepted=%s",
                 summary.selected,
@@ -234,7 +108,7 @@ def main(argv=None):
                 "BAT transform-discovered command started for batch_size=%s",
                 args.batch_size,
             )
-            summary = transform_discovered_listings(batch_size=args.batch_size)
+            summary = bat_pipeline.transform_discovered_listings(batch_size=args.batch_size)
             logger.info(
                 "BAT transform-discovered summary selected=%s transformed_and_loaded=%s transform_failed=%s load_failed=%s",
                 summary.selected,
@@ -257,11 +131,11 @@ def main(argv=None):
 
         logger.info("BAT %s command started for listing_id=%s", args.command, args.listing_id)
         if args.command == "ingest":
-            ingest_listing(args.listing_id)
+            bat_pipeline.ingest_listing(args.listing_id)
         elif args.command == "transform":
-            transform_listing(args.listing_id)
+            bat_pipeline.transform_listing(args.listing_id)
         elif args.command == "run":
-            run_listing(args.listing_id)
+            bat_pipeline.run_listing(args.listing_id)
     except Exception:
         if args.command == "discover":
             logger.error(
